requirements.txt
azure-core==1.32.0
azure-storage-blob==12.25.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
docx2txt==0.9
Flask==3.1.0
Flask-Markdown==0.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
gitingest==0.1.4
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pycparser==2.22
pypdf==5.4.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tiktoken==0.9.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uuid==1.30
Werkzeug==3.1.3
WTForms==3.2.1
